Implement plugins API endpoints - Migrate todos and nuggets API endpoints to their corresponding modules - Implement dynamic fields which can be added to an incident
