+++ fir_api/views.py
@@ -441,130 +441,6 @@ class IncidentCategoriesViewSet(viewsets.ReadOnlyModelViewSet):
     filter_backends = [DjangoFilterBackend, OrderingFilter]
 
 
-if apps.is_installed("fir_todos"):
-    from fir_todos.models import TodoItem
-    from fir_api.serializers import TodoSerializer
-    from fir_api.filters import TodoFilter
-
-    class TodoViewSet(viewsets.ModelViewSet):
-        """
-        API endpoints for listing, creating or editing Todo items
-        """
-
-        queryset = TodoItem.objects.all()
-        serializer_class = TodoSerializer
-        permission_classes = (IsAuthenticated, IsIncidentHandler)
-        filter_backends = [DjangoFilterBackend, OrderingFilter]
-        ordering_fields = ["id", "incident", "done_time"]
-        filterset_class = TodoFilter
-
-        def get_queryset(self):
-            incidents_allowed = Incident.authorization.for_user(
-                self.request.user, "incidents.view_incidents"
-            )
-            queryset = TodoItem.objects.filter(incident__in=incidents_allowed)
-            return queryset
-
-        def perform_create(self, serializer):
-            business_line = BusinessLine.objects.get(
-                name=self.request.data.get("business_line")
-            )
-            if business_line:
-                allowed_businesslines = BusinessLine.authorization.for_user(
-                    self.request.user, "incidents.handle_incidents"
-                )
-                if business_line not in allowed_businesslines:
-                    return Response(
-                        data={"Error": "Chosen businessline not allowed for user"},
-                        status=status.HTTP_403_FORBIDDEN,
-                    )
-            else:
-                request_business_line = None
-            incident_object_id = self.request.data.get("incident")
-            incident_object = Incident.objects.get(pk=incident_object_id)
-            self.check_object_permissions(self.request, incident_object)
-            category = incident_object.category
-            serializer.save(category=category, business_line=business_line),
-
-        def perform_update(self, serializer):
-            self.check_object_permissions(self.request, serializer.instance.incident)
-            business_line = BusinessLine.objects.get(
-                name=self.request.data.get("business_line")
-            )
-            if business_line:
-                allowed_businesslines = BusinessLine.authorization.for_user(
-                    self.request.user, "incidents.handle_incidents"
-                )
-                if business_line not in allowed_businesslines:
-                    return Response(
-                        {"error": "Chosen businessline not allowed for user"},
-                        status=401,
-                    )
-            incident_object_id = self.request.data.get("incident")
-            incident_object = Incident.objects.get(pk=incident_object_id)
-            self.check_object_permissions(self.request, incident_object)
-            category = incident_object.category
-            serializer.save(category=category, business_line=business_line)
-
-        def perform_destroy(self, instance):
-            self.check_object_permissions(self.request, instance.incident)
-            return super().perform_destroy(instance)
-
-
-if apps.is_installed("fir_nuggets"):
-    from fir_nuggets.models import Nugget
-    from fir_api.serializers import NuggetSerializer
-    from fir_api.filters import NuggetFilter
-
-    class NuggetViewSet(viewsets.ModelViewSet):
-        """
-        API endpoints for listing, creating or editing Nuggets
-        """
-
-        queryset = Nugget.objects.all()
-        serializer_class = NuggetSerializer
-        permission_classes = (IsAuthenticated, IsIncidentHandler)
-        filter_backends = [DjangoFilterBackend, OrderingFilter]
-        ordering_fields = [
-            "id",
-            "source",
-            "start_timestamp",
-            "end_timestamp",
-            "interpretation",
-            "incident",
-        ]
-        filterset_class = NuggetFilter
-
-        def get_queryset(self):
-            incidents_allowed = Incident.authorization.for_user(
-                self.request.user, "incidents.view_incidents"
-            )
-            queryset = Nugget.objects.filter(incident__in=incidents_allowed)
-            return queryset
-
-        def perform_create(self, serializer):
-            incident_object_id = self.request.data.get("incident")
-            incident_object = Incident.objects.get(pk=incident_object_id)
-            self.check_object_permissions(self.request, incident_object)
-            instance = serializer.save(found_by=self.request.user)
-            incident_object.refresh_artifacts(instance.raw_data)
-
-        def perform_update(self, serializer):
-            self.check_object_permissions(self.request, serializer.instance.incident)
-            instance = serializer.save()
-            e = get_object_or_404(
-                Incident.authorization.for_user(
-                    self.request.user, "incidents.handle_incidents"
-                ),
-                pk=instance.incident_id,
-            )
-            e.refresh_artifacts(instance.raw_data)
-
-        def perform_destroy(self, instance):
-            self.check_object_permissions(self.request, instance.incident)
-            return super().perform_destroy(instance)
-
-
 # Token Generation ===========================================================
